#88 [기록화면, 이미지 추가화면] 이미지 삭제, 추가 여러개

+++ app/src/main/java/com/thequietz/travelog/data/Repository.kt
@@ -99,12 +99,7 @@ class GuideRepository @Inject constructor(
         pageNo: Int
     ): List<RecommendPlace> {
         return try {
-            val res = guideRecommendService.requestAreaBased(
-                NEW_TOUR_API_KEY,
-                areaCode,
-                requestType,
-                pageNo
-            )
+            val res = guideRecommendService.requestAreaBased(NEW_TOUR_API_KEY, areaCode, requestType, pageNo)
             val maxPage = (res.response.body.totalCnt / 10) + 1
             if (maxPage < pageNo) {
                 emptyRecommendList
@@ -131,12 +126,7 @@ class GuideRepository @Inject constructor(
 */
     suspend fun loadFestivalData(areaCode: String, pageNo: Int): List<RecommendPlace> {
         return try {
-            val res = guideRecommendService.requestFestival(
-                NEW_TOUR_API_KEY,
-                getTodayDate(),
-                areaCode,
-                pageNo
-            )
+            val res = guideRecommendService.requestFestival(NEW_TOUR_API_KEY, getTodayDate(), areaCode, pageNo)
             val maxPage = (res.response.body.totalCnt / 10) + 1
             if (maxPage < pageNo) {
                 emptyRecommendList
@@ -176,7 +166,9 @@ class RecordRepository @Inject constructor(
     fun deleteRecordImage(id: Int) {
         coroutineScope.launch {
             val data = recordImageDao.loadRecordImageById(id)
-            recordImageDao.delete(data)
+            if (data != null) {
+                recordImageDao.delete(data)
+            }
         }
     }
